requirements.txt
asgiref==3.7.2
certifi==2020.6.20
cffi==1.16.0
chardet==3.0.4
cryptography==3.4.7
Django==4.2.7
html5lib==1.1
idna==2.10
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
Pillow==8.1.2
pycparser==2.20
pyOpenSSL==20.0.1
PyPDF2==1.26.0
pytz==2021.1
razorpay==1.2.0
reportlab==3.5.67
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
stripe==2.50.0
urllib3==1.25.10
webencodings==0.5.1
xhtml2pdf==0.2.4
